Make normalization of lyrics re-usable using fn
