+++ src/redux/utils/index.js
@@ -1,0 +1,3 @@
+import _normalizeLyrics from './normalizeLyrics';
+
+export const normalizeLyrics = _normalizeLyrics;
